Rename 'object.py' to 'obj.py'
